refactor: ScheduleTable 및 DraggableSchedule 컴포넌트 구조 변경 및 최적화
