+++ src/ScheduleTable/DraggableSchedule.tsx
@@ -1,0 +1,89 @@
+/**
+ * 드래그 가능한 스케줄 아이템 컴포넌트
+ * 강의 정보를 표시하고 삭제 기능을 제공합니다.
+ */
+import { ComponentProps, memo } from "react";
+import { CellSize, DAY_LABELS } from "../constants.ts";
+import { Schedule } from "../types.ts";
+import {
+  Box,
+  Button,
+  Popover,
+  PopoverArrow,
+  PopoverBody,
+  PopoverCloseButton,
+  PopoverContent,
+  PopoverTrigger,
+  Text,
+} from "@chakra-ui/react";
+import { useDraggable } from "@dnd-kit/core";
+import { CSS } from "@dnd-kit/utilities";
+
+// 단일 강의 아이템
+const DraggableSchedule = memo(
+  ({
+    id,
+    data,
+    bg,
+    onDeleteButtonClick,
+  }: { id: string; data: Schedule } & ComponentProps<typeof Box> & {
+      onDeleteButtonClick: () => void;
+    }) => {
+    const { day, range, room, lecture } = data;
+
+    /**
+     * id - 드래그 아이템 구분
+     * setNodeRef - 드래그할 노드 연결
+     * listeners - 마우스/터치 이벤트 리스너
+     * transform - 현재 드래그 중 이동 위치
+     */
+    const { attributes, setNodeRef, listeners, transform } = useDraggable({
+      id,
+    });
+
+    const leftIndex = DAY_LABELS.indexOf(day as (typeof DAY_LABELS)[number]);
+    const topIndex = range[0] - 1;
+    const size = range.length;
+
+    return (
+      <Popover>
+        <PopoverTrigger>
+          <Box
+            position="absolute"
+            left={`${120 + CellSize.WIDTH * leftIndex + 1}px`}
+            top={`${40 + (topIndex * CellSize.HEIGHT + 1)}px`}
+            width={CellSize.WIDTH - 1 + "px"}
+            height={CellSize.HEIGHT * size - 1 + "px"}
+            bg={bg}
+            p={1}
+            boxSizing="border-box"
+            cursor="pointer"
+            ref={setNodeRef}
+            transform={CSS.Translate.toString(transform)}
+            {...listeners}
+            {...attributes}
+          >
+            <Text fontSize="sm" fontWeight="bold">
+              {lecture.title}
+            </Text>
+            <Text fontSize="xs">{room}</Text>
+          </Box>
+        </PopoverTrigger>
+
+        {/* 강의 삭제 모달 */}
+        <PopoverContent onClick={(event) => event.stopPropagation()}>
+          <PopoverArrow />
+          <PopoverCloseButton />
+          <PopoverBody>
+            <Text>강의를 삭제하시겠습니까?</Text>
+            <Button colorScheme="red" size="xs" onClick={onDeleteButtonClick}>
+              삭제
+            </Button>
+          </PopoverBody>
+        </PopoverContent>
+      </Popover>
+    );
+  }
+);
+
+export default DraggableSchedule;
